# docs/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:8080",
  "basePath": "/",
  "tags": [
    {
      "name": "User",
      "description": "Endpoints"
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "access-token",
      "in": "header"
    },
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "https://petstore.swagger.io/oauth/authorize",
      "flow": "implicit",
      "scopes": {
        "read_pets": "read your pets",
        "write_pets": "modify pets in your account"
      }
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/client/register": {
      "post": {
        "tags": [
          "Client"
        ],
        "description": "client haqida api",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "first_name": {
                  "example": "any"
                },
                "last_name": {
                  "example": "any"
                },
                "father_name": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "name": {
                  "example": "any"
                },
                "url": {
                  "example": "any"
                },
                "phone": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                },
                "role": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário encontrado.",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "400": {
            "description": "Xatolik roy berdi"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/client/list": {
      "post": {
        "tags": [
          "Client"
        ],
        "description": "client haqida api",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "pageNumber": {
                  "example": "any"
                },
                "pageSize": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário encontrado.",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "400": {
            "description": "Xatolik roy berdi"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/client/delete/{id}": {
      "get": {
        "tags": [
          "Client"
        ],
        "description": "client haqida api",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário encontrado.",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "400": {
            "description": "Xatolik roy berdi"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/client/getone/{id}": {
      "get": {
        "tags": [
          "Client"
        ],
        "description": "client haqida api",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário encontrado.",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "400": {
            "description": "Xatolik roy berdi"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/welcome": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "Endpoint para obter um usuário.",
        "parameters": [
          {
            "name": "filtro",
            "in": "query",
            "type": "string",
            "description": "Um filtro qualquer."
          },
          {
            "name": "id",
            "in": "query",
            "type": "string",
            "description": "ID haqida"
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário encontrado.",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "400": {
            "description": "Xatolik roy berdi"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Jhon Doe"
        },
        "age": {
          "type": "number",
          "example": 29
        },
        "parents": {
          "type": "object",
          "properties": {
            "father": {
              "type": "string",
              "example": "Simon Doe"
            },
            "mother": {
              "type": "string",
              "example": "Marie Doe"
            }
          }
        },
        "diplomas": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "school": {
                "type": "string",
                "example": "XYZ University"
              },
              "year": {
                "type": "number",
                "example": 2020
              },
              "completed": {
                "type": "boolean",
                "example": true
              },
              "internship": {
                "type": "object",
                "properties": {
                  "hours": {
                    "type": "number",
                    "example": 290
                  },
                  "location": {
                    "type": "string",
                    "example": "XYZ Company"
                  }
                }
              }
            }
          }
        }
      }
    },
    "AddUser": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Jhon Doe"
        },
        "age": {
          "type": "number",
          "example": 29
        },
        "about": {
          "type": "string",
          "example": ""
        }
      },
      "required": [
        "name",
        "age"
      ]
    }
  }
}